+++ src/Hook/index.tsx
@@ -1,4 +1,5 @@
 import useGeolocation from './useGeolocation';
 import useKakaoLoader from './useKakaoLoader';
+import useMapMarker from './useMapMarker';
 
-export { useGeolocation, useKakaoLoader };
+export { useGeolocation, useKakaoLoader, useMapMarker };
